Codigo/gs_config/code/requirements.txt
functions-framework==3.*
google-cloud-bigquery
gspread